backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-multipart
requests
reportlab
python-dotenv

